ConstantCompressbilityBrinePvt: move implementation to impl file
