+++ opm/material/fluidsystems/blackoilpvt/ConstantCompressibilityBrinePvt.hpp
@@ -29,17 +29,15 @@
 
 #include <opm/material/common/Tabulated1DFunction.hpp>
 
-#if HAVE_ECL_INPUT
-#include <opm/input/eclipse/EclipseState/EclipseState.hpp>
-#include <opm/input/eclipse/Schedule/Schedule.hpp>
-#include <opm/input/eclipse/EclipseState/Tables/PvtwsaltTable.hpp>
-#endif
-
 #include <vector>
 
 namespace Opm {
 template <class Scalar, bool enableThermal, bool enableBrine>
 class WaterPvtMultiplexer;
+
+class EclipseState;
+class Schedule;
+
 /*!
  * \brief This class represents the Pressure-Volume-Temperature relations of the gas phase
  *        without vaporized oil.
@@ -56,77 +54,17 @@ public:
                                     const std::vector<TabulatedFunction> formationVolumeTables,
                                     const std::vector<TabulatedFunction> compressibilityTables,
                                     const std::vector<TabulatedFunction> viscosityTables,
-                                    const std::vector<TabulatedFunction> viscosibilityTables)
-        : waterReferenceDensity_(waterReferenceDensity)
-        , referencePressure_(referencePressure)
-        , formationVolumeTables_(formationVolumeTables)
-        , compressibilityTables_(compressibilityTables)
-        , viscosityTables_(viscosityTables)
-        , viscosibilityTables_(viscosibilityTables)
-    { }
+                                    const std::vector<TabulatedFunction> viscosibilityTables);
+
 #if HAVE_ECL_INPUT
     /*!
      * \brief Sets the pressure-dependent water viscosity and density
      *        using a table stemming from the Eclipse PVTWSALT keyword.
      */
-    void initFromState(const EclipseState& eclState, const Schedule&)
-    {
-        const auto& tableManager = eclState.getTableManager();
-        size_t numRegions = tableManager.getTabdims().getNumPVTTables();
-        const auto& densityTable = tableManager.getDensityTable();
-
-        formationVolumeTables_.resize(numRegions);
-        compressibilityTables_.resize(numRegions);
-        viscosityTables_.resize(numRegions);
-        viscosibilityTables_.resize(numRegions);
-        referencePressure_.resize(numRegions);
-
-        const auto& pvtwsaltTables = tableManager.getPvtwSaltTables();
-        if(!pvtwsaltTables.empty()){
-            assert(numRegions == pvtwsaltTables.size());
-            for (unsigned regionIdx = 0; regionIdx < numRegions; ++ regionIdx) {
-                const auto& pvtwsaltTable = pvtwsaltTables[regionIdx];
-                const auto& c = pvtwsaltTable.getSaltConcentrationColumn();
-
-                const auto& B = pvtwsaltTable.getFormationVolumeFactorColumn();
-                formationVolumeTables_[regionIdx].setXYContainers(c, B);
-
-                const auto& compressibility = pvtwsaltTable.getCompressibilityColumn();
-                compressibilityTables_[regionIdx].setXYContainers(c, compressibility);
-
-                const auto& viscositytable = pvtwsaltTable.getViscosityColumn();
-                viscosityTables_[regionIdx].setXYContainers(c, viscositytable);
-
-                const auto& viscosibility = pvtwsaltTable.getViscosibilityColumn();
-                viscosibilityTables_[regionIdx].setXYContainers(c, viscosibility);
-                referencePressure_[regionIdx] = pvtwsaltTable.getReferencePressureValue();
-            }
-        }
-        else {
-            throw std::runtime_error("PVTWSALT must be specified in BRINE runs\n");
-        }
-
-
-        size_t numPvtwRegions = numRegions;
-        setNumRegions(numPvtwRegions);
-
-        for (unsigned regionIdx = 0; regionIdx < numPvtwRegions; ++ regionIdx) {
-
-            waterReferenceDensity_[regionIdx] = densityTable[regionIdx].water;
-        }
-
-        initEnd();
-    }
+    void initFromState(const EclipseState& eclState, const Schedule&);
 #endif
 
-    void setNumRegions(size_t numRegions)
-    {
-        waterReferenceDensity_.resize(numRegions);
-
-        for (unsigned regionIdx = 0; regionIdx < numRegions; ++regionIdx) {
-            setReferenceDensities(regionIdx, 650.0, 1.0, 1000.0);
-        }
-    }
+    void setNumRegions(size_t numRegions);
 
     /*!
      * \brief Set the water reference density [kg / m^3]
@@ -134,8 +72,7 @@ public:
     void setReferenceDensities(unsigned regionIdx,
                                Scalar /*rhoRefOil*/,
                                Scalar /*rhoRefGas*/,
-                               Scalar rhoRefWater)
-    { waterReferenceDensity_[regionIdx] = rhoRefWater; }
+                               Scalar rhoRefWater);
 
     /*!
      * \brief Finish initializing the water phase PVT properties.
@@ -157,10 +94,7 @@ public:
     Evaluation internalEnergy(unsigned,
                         const Evaluation&,
                         const Evaluation&,
-                        const Evaluation&) const
-    {
-        throw std::runtime_error("Requested the enthalpy of water but the thermal option is not enabled");
-    }
+                        const Evaluation&) const;
 
     /*!
      * \brief Returns the dynamic viscosity [Pa s] of the fluid phase given a set of parameters.
@@ -169,20 +103,7 @@ public:
     Evaluation viscosity(unsigned regionIdx,
                          const Evaluation& temperature,
                          const Evaluation& pressure,
-                         const Evaluation& saltconcentration) const
-    {
-        // cf. ECLiPSE 2013.2 technical description, p. 114
-        Scalar pRef = referencePressure_[regionIdx];
-        const Evaluation C = compressibilityTables_[regionIdx].eval(saltconcentration, /*extrapolate=*/true);
-        const Evaluation Cv = viscosibilityTables_[regionIdx].eval(saltconcentration, /*extrapolate=*/true);
-        const Evaluation BwRef = formationVolumeTables_[regionIdx].eval(saltconcentration, /*extrapolate=*/true);
-        const Evaluation Y = (C-Cv)* (pressure - pRef);
-        Evaluation MuwRef = viscosityTables_[regionIdx].eval(saltconcentration, /*extrapolate=*/true);
-
-        const Evaluation& bw = inverseFormationVolumeFactor(regionIdx, temperature, pressure, saltconcentration);
-
-        return MuwRef*BwRef*bw/(1 + Y*(1 + Y/2));
-    }
+                         const Evaluation& saltconcentration) const;
 
     /*!
      * \brief Returns the formation volume factor [-] of the fluid phase.
@@ -191,17 +112,7 @@ public:
     Evaluation inverseFormationVolumeFactor(unsigned regionIdx,
                                             const Evaluation& /*temperature*/,
                                             const Evaluation& pressure,
-                                            const Evaluation& saltconcentration) const
-    {
-        Scalar pRef = referencePressure_[regionIdx];
-
-        const Evaluation BwRef = formationVolumeTables_[regionIdx].eval(saltconcentration, /*extrapolate=*/true);
-        const Evaluation C = compressibilityTables_[regionIdx].eval(saltconcentration, /*extrapolate=*/true);
-        const Evaluation X = C * (pressure - pRef);
-
-        return (1.0 + X*(1.0 + X/2.0))/BwRef;
-
-    }
+                                            const Evaluation& saltconcentration) const;
 
     const Scalar waterReferenceDensity(unsigned regionIdx) const
     { return waterReferenceDensity_[regionIdx]; }
@@ -221,15 +132,7 @@ public:
     const std::vector<TabulatedFunction>& viscosibilityTables() const
     { return viscosibilityTables_; }
 
-    bool operator==(const ConstantCompressibilityBrinePvt<Scalar>& data) const
-    {
-        return this->waterReferenceDensity_ == data.waterReferenceDensity_ &&
-               this->referencePressure() == data.referencePressure() &&
-               this->formationVolumeTables() == data.formationVolumeTables() &&
-               this->compressibilityTables() == data.compressibilityTables() &&
-               this->viscosityTables() == data.viscosityTables() &&
-               this->viscosibilityTables() == data.viscosibilityTables();
-    }
+    bool operator==(const ConstantCompressibilityBrinePvt<Scalar>& data) const;
 
 private:
     std::vector<Scalar> waterReferenceDensity_;
@@ -238,8 +141,11 @@ private:
     std::vector<TabulatedFunction> compressibilityTables_;
     std::vector<TabulatedFunction> viscosityTables_;
     std::vector<TabulatedFunction> viscosibilityTables_;
+
 };
 
 } // namespace Opm
+
+#include <opm/material/fluidsystems/blackoilpvt/ConstantCompressibilityBrinePvt_impl.hpp>
 
 #endif
